pyproject.toml
[project]
name = "lederband"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "dataclasses-json>=0.6.7",
    "pillow>=12.1.0",
]
